requirements.txt
# Core Django stack
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0
drf-spectacular>=0.27
django-cors-headers>=4.0

# Database / PostGIS
psycopg2-binary>=2.9

# Data / ML stack (Peyton models + web ML core)
numpy>=1.26
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
interpret>=0.6

# Ingestion / validation helpers
PyYAML>=6.0          # for ingestion/config/mmucc_schema.yml
python-magic>=0.4.27 # provides the "magic" module for MIME sniffing
clamd>=1.0.2         # optional, used for ClamAV antivirus scanning

# WSGI server (used by Docker/production)
gunicorn>=21.0

# Windows timezone data (recommended for Django on Windows)
tzdata>=2024.1

# NOTE: The 'xgboost' wheel from PyPI is typically CPU-only on Windows.
# For real GPU acceleration you may need a CUDA-enabled XGBoost build
# installed separately (e.g., via conda); if you do that, you can comment
# out the xgboost line above.
